requirements.txt
langchain==0.0.231
openai==0.27.8
python-dotenv==1.0.0
streamlit==1.25.0
